platform: drivers: update mi-V ihc to version 2 in the ip catalog
Starting from v2024.06 release and onwards, the Icicle Kit reference
design will use the Mi-V IHC IP available in the Libero catalog.
Update the mpfs-rpmsg-bm and mpfs-rpmsg-freertos projects to use the new driver.

The Mi-V IHC IP version 2 is not backwards compatible with the
Mi-V IHC subsystem used in the Icicle Kit reference design 2024.02
or earlier. For this reason the polarfire-soc-amp examples in the v2024.06
release will not be backwards compatible with previous releases.

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>