middleware: rpmsg-lite: update isr counters in bm and freertos environments
The rpmsg_lite uses an array to keep track of registered ISR's. The
default aray size is 12. Increment this counter to support up to 28
channels. The format of this counter resides on the linker_id rathen than
the channel id. The maximum linker_id for Mi-V IHC is defined in rpmsg_platform.h

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>