platform: replace mbadaddr with mtval
mbadaddr is a CSR that was present in older RISC-V privileged
specifications but has been removed in more recent versions and replaced
by mtval. As the RISC-V ecosystem has moved on, update the CSR to
reflect this and update the note to inform users using older compilers
that they can revert this change if getting compilation errors.

Signed-off-by: Jamie Gibbons <jamie.gibbons@microchip.com>